+++ back-end/match/src/main/java/com/ssafy/codearena/MatchApplication.java
@@ -1,4 +1,4 @@
-package com.ssafy.match;
+package com.ssafy.codearena;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
