+++ back-end/match/src/main/java/com/ssafy/codearena/config/JasyptConfig.java
@@ -1,4 +1,4 @@
-package com.ssafy.match;
+package com.ssafy.codearena.config;
 
 
 import org.jasypt.encryption.StringEncryptor;
