+++ back-end/match/src/main/java/com/ssafy/codearena/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.match;
+package com.ssafy.codearena.dto;
 
 
 public class UserDto {
