+++ back-end/match/src/main/java/com/ssafy/codearena/exception/ExpressionException.java
@@ -1,24 +1,32 @@
-package com.ssafy.match;
+package com.ssafy.codearena.exception;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ssafy.codearena.dto.MessageDto;
 import org.springframework.web.socket.TextMessage;
 
-public class MatchPlayerNotFoundException extends Exception {
+public class ExpressionException extends NullPointerException{
+    public ExpressionException() {
+        super();
+    }
+
+    public ExpressionException(String s) {
+        super(s);
+    }
+
     static final ObjectMapper objectMapper = new ObjectMapper();
-    static TextMessage TEXT_MESSAGE ;
-    static MessageDto msg = new MessageDto();
+    public static TextMessage TEXT_MESSAGE ;
+
     static{
         MessageDto messageDto = new MessageDto();
         messageDto.setType(MessageDto.MessageType.EXCEPTION);
-        messageDto.setContent("매치메이킹 플레이어 Not Found Exception");
+        messageDto.setContent("null 데이터가 들어왔습니다. 데이터를 확인해주세요");
         try {
             TEXT_MESSAGE = new TextMessage(objectMapper.writeValueAsString(messageDto));
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
         }
     }
-
     TextMessage getTextMessage(){
         return TEXT_MESSAGE;
     }
