+++ back-end/match/src/main/java/com/ssafy/codearena/exception/MatchNotFoundException.java
@@ -1,32 +1,25 @@
-package com.ssafy.match;
+package com.ssafy.codearena.exception;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ssafy.codearena.dto.MessageDto;
 import org.springframework.web.socket.TextMessage;
 
-public class ExpressionException extends NullPointerException{
-    public ExpressionException() {
-        super();
-    }
-
-    public ExpressionException(String s) {
-        super(s);
-    }
-
+public class MatchNotFoundException  extends Exception {
     static final ObjectMapper objectMapper = new ObjectMapper();
     static TextMessage TEXT_MESSAGE ;
 
     static{
         MessageDto messageDto = new MessageDto();
         messageDto.setType(MessageDto.MessageType.EXCEPTION);
-        messageDto.setContent("null 데이터가 들어왔습니다. 데이터를 확인해주세요");
+        messageDto.setContent("매치 Not Found Exception");
         try {
             TEXT_MESSAGE = new TextMessage(objectMapper.writeValueAsString(messageDto));
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
         }
     }
-    TextMessage getTextMessage(){
+    public TextMessage getTextMessage(){
         return TEXT_MESSAGE;
     }
 }
