--- back-end/match/src/main/java/com/ssafy/match/CommonMsg.java
@@ -1,14 +1,0 @@
-package com.ssafy.match;
-
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class CommonMsg {
-    String msg;
-    String matchId;
-}
