--- back-end/match/src/main/java/com/ssafy/match/MyService.java
@@ -1,9 +1,0 @@
-package com.ssafy.match;
-
-import java.util.HashMap;
-import java.util.List;
-
-public interface MyService {
-    String getRedis(RedisDto redisDto);
-    HashMap<String, List<Object>> enqueue(UserDto userDto);
-}
