--- back-end/match/src/main/java/com/ssafy/match/RedisDto.java
@@ -1,30 +1,0 @@
-package com.ssafy.match;
-
-public class RedisDto {
-    private String key;
-    private String value;
-
-    public String getKey() {
-        return key;
-    }
-
-    public void setKey(String key) {
-        this.key = key;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    @Override
-    public String toString() {
-        return "RedisDto{" +
-                "key='" + key + '\'' +
-                ", value='" + value + '\'' +
-                '}';
-    }
-}
