--- back-end/match/src/main/java/com/ssafy/match/YesOrNoDto.java
@@ -1,8 +1,0 @@
-package com.ssafy.match;
-
-public class YesOrNoDto {
-    public enum MessageType{
-        YES, NO
-    }
-    private MessageType type;
-}
