+++ back-end/match/src/main/resources/application.properties
@@ -7,4 +7,4 @@ redis.password=ENC(NkkYJIp6COsEFz7vT99z74UHriiQGaKK)
 
 gameserver.url=ENC(1BY2vKjHkgiXU8K3RwEFydgQxh611KCVkuc2I/eUySM=)
 restserver.url=ENC(g98wNdmJfcVgvEhW3nEBoivUhKBs8jVA7nhOp+HgeTk=)
-logging.level.com.ssafy.match=debug
+logging.level.com.ssafy.codearena=debug
