+++ back-end/match/src/test/java/com/ssafy/codearena/MatchApplicationTests.java
@@ -1,7 +1,6 @@
-package com.ssafy.match;
+package com.ssafy.codearena;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
 class MatchApplicationTests {
